Refactor: Move watch later and add playlist options to a dropdown menu
This commit refactors the watch later button and adds an "Add to Playlist" option within a dropdown menu. This change improves the user interface by decluttering the floating controls and providing a more organized way to manage video options. The dropdown menu is implemented using `AnimatePresence` for smooth transitions. Also added download and share options to the dropdown.