Fix: Get user playlists based on logged-in user
This commit fixes an issue where playlists were not being fetched correctly for the logged-in user. The backend was updated to retrieve playlists based on the `req.user._id` instead of `req.params.userID`. The frontend `usePlaylist` hook was also updated to pass the user object. This ensures that only the playlists belonging to the authenticated user are displayed.